refactor : 클래스 및 파일 위치 변경
