+++ src/main/java/com/example/ossdoc/domain/cluster/artifact/output/RankingsJson.java
@@ -1,4 +1,4 @@
-package com.example.ossdoc.domain.cluster.dto.output;
+package com.example.ossdoc.domain.cluster.artifact.output;
 
 import com.example.ossdoc.domain.cluster.model.ranking.SubsystemRankingItem;
 import com.example.ossdoc.domain.cluster.model.ranking.SymbolRankingItem;
@@ -10,7 +10,7 @@ import java.util.List;
 
 @Getter
 @Builder
-public class RankingsJsonDto {
+public class RankingsJson {
     private String schemaVersion;
     private String runId;
     private OffsetDateTime generatedAt;
